configure.ac: fix configure generation with autotools-1.14
Use of "dnl" as the last item on the line removes "newline".
And in case of invocation of macros this will lead to placing of
multiple macros on the same line. This in its turn may break normal
behavior of some macros.

In particular this line is important:
 --->8---
 AM_INIT_AUTOMAKE(1.9.6 foreign dejagnu no-installinfo dist-bzip2)dnl
 --->8---

With "dnl" in the very end of the macro autotools 1.14 produces
incorrect
"configure" script:
 --->8---
 $ ./configure
 checking for a BSD-compatible install... /usr/bin/install -c
 checking whether build environment is sane... yes
 checking for a thread-safe mkdir -p... /usr/bin/mkdir -p
 checking for gawk... gawk
 checking whether make sets $(MAKE)... yes
 checking whether make supports nested variables... yes
 ./configure: line 11561: syntax error: unexpected end of file
 --->8---

Solution is simple - remove trailing "dnl":
 --->8---
 AM_INIT_AUTOMAKE(1.9.6 foreign dejagnu no-installinfo dist-bzip2)
 --->8---

See discussion of this issue here as well:
https://bugzilla.redhat.com/show_bug.cgi?id=1191007

And while at it I removed most of trailing "dnl" as a clean-up in places
that are not related to comments.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>